requirements.txt
pyspark
pyarrow
pandas
numpy
pyyaml
scikit-learn
matplotlib
seaborn
jupyter
